requirements.txt
streamlit
pillow
pytesseract
openai
scikit-learn
joblib
pandas
